scraper/requirements.txt

beautifulsoup4>=4.12.2
requests>=2.31.0
feedparser>=6.0.10
python-dotenv>=1.0.0
pydantic>=2.5.0
lxml>=4.9.3
packaging>=21.0
python-dateutil>=2.8.0
anthropic>=0.39.0